Take belt thickness into account when calculating esteps
The spacing on the GT2 belts is between the belt grooves. The rubber teeth on the belt add to the radius of the belt, causing the effective circumference to larger. This was a problem on the icepick delta until I added these configuration options.

Not sure if you already factored this in with the pulley radius, in which case this addition maybe unnecessary.